requirements.txt
Flask 
azure-data-tables
python-dotenv
flask-cors
azure-storage-blob
openai